requirements.txt
# AI-Essentials Python Dependencies
# Python 3.10+ required

# Core ML/AI libraries
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0

# LLMs and Transformers
openai>=1.0.0
transformers>=4.30.0
torch>=2.0.0

# LangChain for RAG and Agents
langchain>=0.1.0
langchain-openai>=0.0.5

# Vector Databases
chromadb>=0.4.0
pinecone-client>=3.0.0

# RAG Components
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# Agent Frameworks
crewai>=0.1.0

# Jupyter for notebooks
jupyter>=1.0.0
ipykernel>=6.25.0

# MCP (Model Context Protocol)
mcp>=0.1.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
